repr: start datetime module
Start a new datetime module in repr. This reverses the dependency edge
between repr and sql-parser, so that sql-parser properly depends on repr
and not vice-versa. This means that sql-parser can produce
repr::datetime::Intervals directly, which reduces some boilerplate.

Fix #1006.
