+++ src/repr/Cargo.toml
@@ -22,7 +22,6 @@ pretty = "0.7.1"
 regex = "1.3.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_regex = "0.4.0"
-sql-parser = { path = "../sql-parser" }
 
 [dev-dependencies]
 criterion = "0.3"
