+++ src/sql-parser/src/ast/mod.rs
@@ -58,8 +58,7 @@ pub use self::query::{
     SetExpr, SetOperator, TableAlias, TableFactor, TableWithJoins, Values,
 };
 pub use self::value::{
-    DateTimeField, ExtractField, Interval, IntervalValue, ParsedDate, ParsedDateTime,
-    ParsedTimestamp, Value,
+    DateTimeField, ExtractField, IntervalValue, ParsedDate, ParsedDateTime, ParsedTimestamp, Value,
 };
 
 struct DisplaySeparated<'a, T>
